requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
groq==0.11.0
pdfplumber==0.11.4
pymupdf==1.25.5
python-multipart==0.0.20
python-docx==1.1.2
httpx==0.27.2
sqlalchemy==2.0.36
aiosqlite==0.20.0